requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
psycopg2==2.7.7
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
